Move busymachines.result package to busymachines.effects.result
