+++ future/src/main/scala/busymachines/future/futureOpsSyntax.scala
@@ -3,7 +3,7 @@ package busymachines.future
 import busymachines.core.Anomaly
 import busymachines.duration
 import busymachines.duration.FiniteDuration
-import busymachines.result._
+import busymachines.effects.result._
 
 import scala.collection.generic.CanBuildFrom
