+++ result/src/main/scala/busymachines/effects/result/package.scala
@@ -1,4 +1,4 @@
-package busymachines
+package busymachines.effects
 
 /**
   *
