+++ result/src/main/scala/busymachines/effects/result/resultFromOpsSyntax.scala
@@ -1,4 +1,4 @@
-package busymachines.result
+package busymachines.effects.result
 
 import busymachines.core.Anomaly
 import scala.util.Try
